menambahkan beberapa file dan folder

+++ views/list-anggota.php
@@ -23,52 +23,42 @@ $users = Anggota::get();
 
 <body class="sb-nav-fixed">
     <nav class="sb-topnav navbar navbar-expand navbar-dark bg-dark">
-        <a class="navbar-brand ps-3" href="dashboard.php">project01</a>
-        <button class="btn btn-link btn-sm order-1 order-lg-0 me-4 me-lg-0" id="sidebarToggle"><i class="fas fa-bars"></i></button>
-        <form class="d-none d-md-inline-block form-inline ms-auto me-0 me-md-3 my-2 my-md-0">
-            <div class="input-group">
-                <input class="form-control" type="text" placeholder="Search for..." aria-label="Search for..." aria-describedby="btnNavbarSearch" />
-                <button class="btn btn-primary" id="btnNavbarSearch" type="button"><i class="fas fa-search"></i></button>
-            </div>
-        </form>
-        <ul class="navbar-nav ms-auto ms-md-0 me-3 me-lg-4">
-            <li class="nav-item dropdown">
-                <a class="nav-link dropdown-toggle" id="navbarDropdown" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false"><i class="fas fa-user fa-fw"></i></a>
-                <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
-                    <li><a class="dropdown-item" href="#!">Settings</a></li>
-                    <li><a class="dropdown-item" href="#!">Activity Log</a></li>
-                    <li>
-                        <hr class="dropdown-divider" />
-                    </li>
-                    <li><a class="dropdown-item" href="#!">Logout</a></li>
-                </ul>
-            </li>
-        </ul>
+        <a class="navbar-brand ps-3" href="dashboard.php">Koperasi Pegawai</a>
+        <button class="btn btn-link btn-sm order-1 order-lg-0 me-4 me-lg-0" id="sidebarToggle">
+            <i class="fas fa-bars"></i>
+        </button>
     </nav>
+
     <div id="layoutSidenav">
         <div id="layoutSidenav_nav">
             <nav class="sb-sidenav accordion sb-sidenav-dark" id="sidenavAccordion">
                 <div class="sb-sidenav-menu">
                     <div class="nav">
                         <div class="sb-sidenav-menu-heading">Main Menu</div>
-                        <a class="nav-link" href="list-anggota.php">
-                            <div class="sb-nav-link-icon"><i class="fa-solid fa-users"></i></div>
-                            Anggota
+
+                        <a class="nav-link collapsed" href="#" data-bs-toggle="collapse" data-bs-target="#collapseAnggota" aria-expanded="false" aria-controls="collapseAnggota">
+                            <div class="sb-nav-link-icon"><i class="fas fa-tachometer-alt"></i></div>
+                            Manajemen Anggota
+                            <div class="sb-sidenav-collapse-arrow"><i class="fas fa-angle-down"></i></div>
                         </a>
-                        <a class="nav-link" href="list-pegawai.php">
-                            <div class="sb-nav-link-icon"><i class="fa-solid fa-user-tie"></i></div>
-                            Pegawai
-                        </a>
+                        <div class="collapse" id="collapseAnggota" data-bs-parent="#sidenavAccordion">
+                            <nav class="sb-sidenav-menu-nested nav">
+                                <a class="nav-link" href="list-anggota.php">Data Anggota</a>
+                                <a class="nav-link" href="list-pegawai.php">Data Pegawai</a>
+                                <a class="nav-link" href="list-kartu_diskon.php">Kartu Diskon</a>
+                            </nav>
+                        </div>
+
                         <a class="nav-link" href="list-produk.php">
-                            <div class="sb-nav-link-icon"><i class="fa-solid fa-box"></i></div>
+                            <div class="sb-nav-link-icon"><i class="fas fa-box"></i></div>
                             Produk
                         </a>
                         <a class="nav-link" href="list-pesanan.php">
-                            <div class="sb-nav-link-icon"><i class="fa-solid fa-cart-shopping"></i></div>
+                            <div class="sb-nav-link-icon"><i class="fas fa-cart-shopping"></i></div>
                             Pesanan
                         </a>
                         <a class="nav-link" href="list-pembayaran.php">
-                            <div class="sb-nav-link-icon"><i class="fa-solid fa-money-bill"></i></div>
+                            <div class="sb-nav-link-icon"><i class="fas fa-money-bill"></i></div>
                             Pembayaran
                         </a>
                     </div>
@@ -79,6 +69,7 @@ $users = Anggota::get();
                 </div>
             </nav>
         </div>
+
         <div id="layoutSidenav_content">
             <main>
                 <div class="container-fluid px-4">
@@ -105,26 +96,26 @@ $users = Anggota::get();
                                         <th>Status Aktif</th>
                                         <th>Pegawai ID</th>
                                         <th>Kartu Diskon ID</th>
+                                        <th>Action</th>
                                     </tr>
                                 </thead>
                                 <tbody>
                                     <?php foreach ($users as $index => $user) : ?>
                                         <tr>
-                                            <td><?= $user['id'] ?></td>
-                                            <td><?= $user['status_aktif'] ?></td>
-                                            <td><?= $user['pegawai_id'] ?></td>
-                                            <td><?= $user['kartu_diskon_id'] ?></td>
+                                            <td><?= htmlspecialchars($user['id']) ?></td>
+                                            <td><?= $user['status_aktif'] == 1 ? 'Aktif' : 'Tidak Aktif' ?></td>
+                                            <td><?= htmlspecialchars($user['pegawai_id']) ?></td>
+                                            <td><?= htmlspecialchars($user['kartu_diskon_id']) ?></td>
+
                                             <td>
-                                                <a href="detail-anggota.php?id=<?= $user['id'] ?>" class="btn btn-primary">
+                                                <a href="detail-anggota.php?id=<?= $user['id'] ?>" class="btn btn-primary btn-sm">
                                                     <i class="fas fa-eye"></i> Detail
                                                 </a>
-
-                                                <a href="edit-anggota.php?id=<?= $user['id'] ?>" class="btn btn-warning">
-                                                    <i class="fas fa-edit"></i> Detail
+                                                <a href="edit-anggota.php?id=<?= $user['id'] ?>" class="btn btn-warning btn-sm">
+                                                    <i class="fas fa-edit"></i> Edit
                                                 </a>
-
-                                                <a href="delete-anggota.php?id=<?= $user['id'] ?>" class="btn btn-danger">
-                                                    <i class="fas fa-trash"></i> Detail
+                                                <a href="delete-anggota.php?id=<?= $user['id'] ?>" class="btn btn-danger btn-sm" onclick="return confirm('Yakin ingin menghapus anggota ini?');">
+                                                    <i class="fas fa-trash"></i> Delete
                                                 </a>
                                             </td>
                                         </tr>
